refactor: separata state and district custom typse, rename mapProps to map
