+++ src/components/Header/SearchBar/SearchBar.tsx
@@ -5,7 +5,7 @@ import { useDebounce } from 'use-debounce';
 import { Search, Close } from '@mui/icons-material';
 
 import { useSelectedDistrict } from '@store/district/selectedContext';
-import { District } from '@customTypes/feature';
+import { District } from '@customTypes/district';
 
 import useSearch from './hook/useSearch';
