+++ src/components/Header/SearchBar/utils.ts
@@ -1,9 +1,9 @@
-import { Feature } from '@customTypes/feature';
+import { District } from '@customTypes/district';
 
-export function getSortedDistricts(features: Feature[]) {
-  return features?.sort((a: any, b: any) => a?.properties.NM_MUN.localeCompare(b?.properties.NM_MUN));
+export function getSortedDistricts(districts: District[]) {
+  return districts?.sort((a: any, b: any) => a?.properties.NM_MUN.localeCompare(b?.properties.NM_MUN));
 }
 
-export function getFilteredDistricts(features: Feature[], query: string) {
-  return features?.filter((item: any) => item?.properties.NM_MUN.toLowerCase().indexOf(query.toLowerCase()) !== -1);
+export function getFilteredDistricts(districts: District[], query: string) {
+  return districts?.filter((item: any) => item?.properties.NM_MUN.toLowerCase().indexOf(query.toLowerCase()) !== -1);
 }
