+++ src/components/Sidebar/RegionDetails/DataSection.tsx
@@ -7,7 +7,7 @@ import MetricDetails from '@components/MetricDetails';
 import { useSelectedDistrict } from '@store/district/selectedContext';
 import { useComparison } from '@store/comparisonContext';
 
-import { MapPropsContentType, MapPropsSectionType } from '@customTypes/mapProps';
+import { MapPropsContentType, MapPropsSectionType } from '@customTypes/map';
 
 import { Tooltip } from '@mui/material';
