+++ src/config/district/social.ts
@@ -1,5 +1,5 @@
 import { formatValue } from '@utils/formatValue';
-import { MapPropsSectionType } from '@customTypes/mapProps';
+import { MapPropsSectionType } from '@customTypes/map';
 
 const socialProps: MapPropsSectionType = {
   title: 'Social (S)',
