+++ src/customTypes/state.ts
@@ -1,0 +1,18 @@
+export interface State {
+    type: string;
+    _geometry?: {
+        type: string;
+        coordinates: number[][][];
+    };
+    geometry: {
+        type: string;
+        coordinates: number[][][];
+    };
+    properties: {
+        NM_UF?: string;
+        CD_UF?: number;
+        SIGLA_UF: string;
+        AREA_KM2: number;
+        POPULATION?: number;
+    };
+}
