+++ src/store/state/selectedContext.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useContext, useState } from 'react';
 
-import { State } from '@customTypes/feature';
+import { State } from '@customTypes/state';
 
 import geojsonBR from '@data/BR_UF_2020.json';
